+++ src/main/java/com/burgosmanuel/portfolio/persona/IPersonaService.java
@@ -1,7 +1,7 @@
 package com.burgosmanuel.portfolio.persona;
 
 public interface IPersonaService {
-    public Persona buscarPersona(int id);
+    public Persona buscarPersona(Long id);
     public void crearPersona(Persona pers);
-    public void editarPersona(int id, Persona datosPersona);
+    public void editarPersona(Long id, Persona datosPersona);
 }
